delete user moved to editprofile page
