+++ src/pages/UserProfile.js
@@ -3,7 +3,6 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { API_URL } from 'utils/BackendUrl';
 import plants from 'reducers/plants';
 import user from 'reducers/user';
 
@@ -42,23 +41,7 @@ export const UserProfile = () => {
   const onToDoClick = () => {
     navigate(`/${username}/toDo`);
   };
-  const onDeleteUserClick = () => {
-    const options = {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: accessToken
-      }
-    }
-    // eslint-disable-next-line space-unary-ops
-    fetch(API_URL(`${username}`), options)
-      .then((response) => response.json())
-      .then((data) => {
-        console.log(data);
-        dispatch(user.actions.deleteUser());
-        navigate('/')
-      });
-  };
+
   return (
     <div>
       <img className="profile-picture" src={imageUrl} alt="profile" />
@@ -78,10 +61,6 @@ export const UserProfile = () => {
       <button
         type="button"
         onClick={onEditUserClick}> Edit User
-      </button>
-      <button
-        type="button"
-        onClick={onDeleteUserClick}> Delete User
       </button>
     </div>
   )
